+++ web-manager/static/js/modules/camera.js
@@ -1,6 +1,6 @@
 /**
  * RTSP Recorder Web Manager - Preview and camera controls
- * Version: 2.32.92
+ * Version: 2.32.93
  */
 
 (function () {
@@ -1577,4 +1577,5 @@ async function deleteProfile(profileId) {
     window.applyProfile = applyProfile;
     window.deleteProfile = deleteProfile;
 })();
+