+++ web-manager/static/js/modules/diagnostics.js
@@ -1,6 +1,6 @@
 /**
  * RTSP Recorder Web Manager - Diagnostic functions
- * Version: 2.32.92
+ * Version: 2.32.93
  */
 
 (function () {
@@ -1779,4 +1779,5 @@ async function deleteProfile(profileId) {
     window.applyProfile = applyProfile;
     window.deleteProfile = deleteProfile;
 })();
+